Fix wrong network conversion field type (#3796)

+++ blockchainApi/src/main/java/com/blockchain/api/blockchainCard/data/BlockchainCardDto.kt
@@ -82,7 +82,7 @@ data class BlockchainCardTransactionDto(
     val clearedFundingAmount: PriceDto,
     val userTransactionTime: String,
     val merchantName: String,
-    val networkConversionRate: Int?,
+    val networkConversionRate: Float?,
     val declineReason: String?,
     val fee: PriceDto,
 )
